Clean up, added comments, changed icons
