+++ assets/css/mtag.bpmnmodeler.font.css
@@ -159,5 +159,5 @@
 .bpmn-icon-hand-tool:before { content: '\e868'; } /* '' */
 .bpmn-icon-group:before { content: '\e869'; } /* '' */
 .bpmn-icon-transaction:before { content: '\e8c4'; } /* '' */
-.bpmn-icon-undo:before { content: '\2B6F'; }
-.bpmn-icon-redo:before { content: '\2B6E'; }
+.bpmn-icon-undo:before { content: '\2B8C'; }
+.bpmn-icon-redo:before { content: '\2B8C'; }
