+++ demo.html
@@ -64,16 +64,6 @@
   }
 });
 
-// var commandStack = myModeler.get("commandStack");
-
-// document.addEventListener("click", function(event) {
-//   if (event.target.id == 'undo') {
-//     commandStack.undo();
-//   } else if (event.target.id == 'redo') {
-//     commandStack.redo();
-//   }
-// });
-
 async function init() {
 
   openDiagram();
